Fix a bug causing keyboard view offset to be incorrect (#204)
* Fix a bug causing keyboard view offset to be incorrect

In some rare cases, if the height of the view is a fractional point
(i.e., not a whole number), it would cause the views to not be offset
despite the keyboard being shown on top of them. This does not happen
with every fractional height. Different devices also behave a little
differently in seemingly identical layouts, due to their pixel density
being different.

The base issue is that, due to floating point rounding errors, two
values that _should_ be identical and pass the guard fail to do so,
because the lack of precision results in them not being equal. By
flooring the values, we can ignore really minor differences and ensure
rounding errors don't cause this issue.

* Unify bma_round methods to use correct calculation

Thanks to @diegosanchezr for the suggested improvement.

* Revert ChattoAdditions bma_round change

Unfortunately, removing this in favor of the Chatto version broke size
calculations, so putting it back...

* Switch to using infix operator to check float comparison

* Add utils to Chatto project