Fix iOS 11 text rendering issue
allowsNonContiguousLayout should only be used for large documents, and, if so, the ensureGlyph(...) functions must be called. Since we don't deal with large documents, we don't need this option.
An optimization on iOS 11 made the text not to render on some circumstances (for instance when there is a carriage return).